Update patch set 1
Patch Set 1:

> LF is planning to move from storing per-job archive data on nexus to AWS/S3, but it has not been put into production yet.  Hopefully this will reduce the upload time, but it doesn't explain why it would take longer on AARCH64 hosts.  Is there a difference in network bandwidth available on AARCH64 wrt. X86_64 Nomad hosts?

There isn't as far as I know. What does the Robot plugin do when it's "Copying log files to build dir"?

But I also had a different observation: In the past in a regular Jenkins job (non-sandbox, timed) it took a lot less time - just a minute compared to the current 10 minutes in Sandbox.

Patch-set: 1