Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> Not sure how important the logs are. Adding Dave to decide.

This is meant to be a temporary measure until we get the publisher macro fixed.  In the meantime, we verified that the console logs are available in Jenkins (all the way back to job #1), so I'm good with this as a workaround.

Patch-set: 2
Reviewer: Gerrit User 10 <10@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>