run ASAN within the non-voting VPP debug verify job
Address sanitizer catches a wide range of failures
like buffer overruns, etc. So makes sense to run it
during the debug test run.

However, it is a bit picky to the compiler -
currently it only works well only on GCC 8+,
so this patch accounts for that.

Thanks to Benoit for the help getting it working.

Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>
Change-Id: If086907d1f20e99503aa0e811fd2ae95c168d731