delete the references to packagecloud.io .deb repositories after installing vpp-ext-deps
If a new empty packagecloud.io/fdio/XXYY empty repository is being added during the stable/XXYY
branch creation process, the jobs will fail unless a dummy artifact is added to this repository.
Since we only use this repository to install the right version of vpp-ext-deps, delete
the references after that portion is complete.

Change-Id: Ia5dfa5bf0193c0af8671d561d27f9f02c4e641f4
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>