Update patch set 4
Patch Set 4:

So I left the "force the nameserver entry" for now, and encapsulated the Python script as a here-doc, because the shell script becomes part of the XML job description for Jenkins - and I don't see a sane way to add an "external asset". In doing that I had to de-pythonize the f"{foo}" string syntax since it made JJB unhappy since it thought it was a parameter.

Since I create the script on the fly, I use its existence to assert that there was already an attempt to run this shell script - which will happen it is referenced both from legacy and new publishers...

There is also quirkiness in that include-raw-deploy-archives.sh zips and cleans up the logs - subsequently, while the scripts/post_build_deploy_archives.sh merely prepares the $WORKSPACE/archives with the necessary files.

So there is a conditional that checks for the existence of the zipfile and if it is there, it uses it as a source for the logs to be uploaded, or if not - then it copies the data from the archives directory.

to keep itself as transparent as possible it does all the copy operation via a temp location, which I do not clean up since I assume the execution is happening ephemerally. Should I do that ?

Patch-set: 4
Reviewer: Gerrit User 320 <320@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>