Update patch set 5
Patch Set 5:

Looking at [0], I see the files have been uploaded. Similarly to the other publishers, Sandbox is distingushed by "vex-yul-rot-jenkins-2" in the URL (as opposed to *-1 for production Jenkins).
I am working from a corporate VPN, and it really does not trust the *.consul host name. Is there a documentation on how to access the uploaded files from public in a secure way?

Also, we should add some automated cleanup for logs from Sandbox, probably into this [1] job.
After a while (if/when we decide the new location should be primary, not just a backup), we can switch the description setter so use the new URL.

[0] https://jenkins.fd.io/sandbox/job/vpp-csit-verify-api-crc-master/1/console
[1] https://jenkins.fd.io/view/ci-management/job/ci-management-jenkins-sandbox-cleanup/

Patch-set: 5
CC: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>