Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Looking at [0], I see the files have been uploaded. Similarly to the other publishers, Sandbox is distingushed by "vex-yul-rot-jenkins-2" in the URL (as opposed to *-1 for production Jenkins).

The Path is different. It is a problem of bash wrapping of path (moving to proper dir). For not not a big deal unless we know how AWS S3 operates.

> I am working from a corporate VPN, and it really does not trust the *.consul host name. Is there a documentation on how to access the uploaded files from public in a secure way?
> 

The way forward is to expose service to public. Once we agree on whether we want to go public or not.

> Also, we should add some automated cleanup for logs from Sandbox, probably into this [1] job.
> After a while (if/when we decide the new location should be primary, not just a backup), we can switch the description setter so use the new URL.

It is just a matter of policy which right now assumes logs rotation in 180d. As sandbox and production is one bucket that will not be possible. But we might still want to create sandbox bucket with own rules.

Patch-set: 5
Reviewer: Gerrit User 76 <76@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>