Modify OpenStack templates
Switching the OpenStack templates used by CI mgmt jobs back to V2
flavors due to problems with the V3 flavors in production. Modifying the
INFO.yaml job to use a template with a V3 flavor to continue
troubleshooting the issue in production because this issue can't be
reproduced on the sandbox.

Signed-off-by: Vanessa Rene Valderrama <vvalderrama@linuxfoundation.org>
Change-Id: Ibc8e744e3201538122965a70ab0f23ca3fe66371