Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> We already had naginator and it was doing more hard then good with voting. Is this fixed?

In at least the following case, the job which failed [0] and was retried [1] occurred quickly and did not disrupt the voting for the patch [2].  

Unfortunately the failure & retry did not get recorded in gerrit log, so the retry was only detected by looking at each job queue and reviewing all failures.

The good news is that there was only a single retry so there was no voting anomaly in gerrit.  Hopefully this will continue be true, but there is still a potential to get unlucky.

I will continue to monitor failures/retries of all VPP verify/merge jobs on master.

[0] https://jenkins.fd.io/job/vpp-verify-master-centos8-x86_64/3335/
[1] https://jenkins.fd.io/job/vpp-verify-master-centos8-x86_64/3336/
[2] https://gerrit.fd.io/r/c/vpp/+/32206

Patch-set: 1