Update patch set 5
Patch Set 5: Code-Review-1

> Just use the existing 2n-clx for testing, such as testing in a sandbox

I see the job, and it is OK to run the tests on 2n-clx while developing. But the agreement on CSIT weekly meeting was to use "device" testbeds, not "perf" testbeds such as 2n-clx.

The device testbeds have different topology (1-node topology, meaning no traffic generator node, just the SUT with cable looping back to a different port of the same NIC), and reservation (in order to run multiple tests at once, we use SR-IOV and a script to assign different VF pair for each run) so I half-expect the CSIT code will need to be updated (or maybe not if it happens to detect everything properly).
I think we still need to install Intel-E810CQ into 1n-skx device SUT and update the VF reservation scripts to reserve VFs on it. Then the flow suites can be moved to tests/vpp/device so they are running together wit the existing device tests. With that plan, no ci-management changes are needed.

We can discuss on the next CSIT weekly call.

Patch-set: 5
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=-1