Description setter: Link to S3 log archive only
The description-setter plugin [0] has a limitation:
"This RegEx is used to scan the build log file,
line by line (multi lines are not supported),
and it will take the first line that matches your RegEx."

Currently, we archive logs both into Nexus and S3,
but the order is hardcoded in global-jjb [1],
so the current RegEx will always match the Nexus line.

As we move away from Nexus, we want the S3 link
to be the primary source of logs, but I do not see
an easy way to configure that in global-jjb
(there may be many projects using the scipt
which are not ready for their RegEx suddenly matching
the S3 link, especially if their S3 upload fails).

This Change affects only fd.io jobs
by simply not searching for the Nexus line,
thus always showing the S3 line.

[0] https://plugins.jenkins.io/description-setter/
[1] https://github.com/lfit/releng-global-jjb/blob/92449c24828b0dbd582637ce3d1d705390da4126/shell/logs-deploy.sh

Change-Id: I35961624b5fe68e5d41f77d02a8cff9d5b2af6bd
Signed-off-by: Vratko Polak <vrpolak@cisco.com>