Update patch set 1
Patch Set 1:

> 32650. This should resolve that issue.

That adds a new macro, which is not used anywhere (so not in the udpi jobs we are discussing here).

Can you create a complete solution for updi jobs (e.g. create a chain of Gerrit changes, or squash all the edits here) so we can test it on Sandbox?
Just from this patch and 32650 I have trouble envisioning which commands will be executed after all the edits.

Patch-set: 1
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>