Update patch set 4
Patch Set 4: Code-Review-2

Robot outputs need to be archived in a place available publicly (consul is not).
It is possible the LFN S3 link will work when Nexus upload fails, but I need to see a Sandbox run with a non-nexus link in run description that opens when I click on it (with log.html somewhere inside).

Is there a way to test this without affecting all Sandbox jobs, for example unsetting the variables in bash just for one job?

Patch-set: 4
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=-2