Update patch set 2
Patch Set 2:

It seems my idea of a run keeping its configuration when the job is reconfigured when the run is in progress is wrong, and the final configuration failed due to 401 on Nexus upload attempt.
https://jenkins.fd.io/sandbox/job/csit-vpp-perf-verify-master-2n-skx/10/console

So this will take a while to verify on Sandbox properly. If not today then next week.

Patch-set: 2
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>